+++ profile/src/main/java/com/bank/profile/dto/RegistrationDto.java
@@ -5,16 +5,13 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-import java.time.LocalDate;
-import java.util.Set;
 
 @AllArgsConstructor
 @Setter
 @Getter
 @Builder
 public class RegistrationDto {
-    private BigInteger id;
+    private Long id;
     private String country;
     private String region;
     private String city;
