+++ profile/src/main/java/com/bank/profile/service/ProfileService.java
@@ -2,19 +2,18 @@ package com.bank.profile.service;
 
 import com.bank.profile.dto.ProfileDto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ProfileService {
     public List<ProfileDto> findAll();
 
-    public ProfileDto findOne(BigInteger id);
+    public ProfileDto findOne(Long id);
 
     public ProfileDto findByInnOrPhoneNumber(Long n);
 
     public ProfileDto create(ProfileDto profileDto);
 
-    public boolean delete(BigInteger id);
+    public boolean delete(Long id);
 
     public ProfileDto update(ProfileDto profileDto);
 }
